requirements.txt

streamlit
pandas
numpy
plotly
folium
Pillow
rasterio
streamlit-folium
